+++ examples/outstanding-instance/main.tf
@@ -3,13 +3,9 @@ provider "ucloud" {
   region = var.region
 }
 
-# Query availability zone
-data "ucloud_zones" "default" {
-}
-
 # Query image
 data "ucloud_images" "default" {
-  availability_zone = data.ucloud_zones.default.zones[1].id
+  availability_zone = var.zone
   name_regex        = "^高内核CentOS 7.6 64位"
   image_type        = "base"
 }
@@ -21,7 +17,7 @@ data "ucloud_security_groups" "default" {
 
 # Create a outstanding instance
 resource "ucloud_instance" "outstanding" {
-  availability_zone = data.ucloud_zones.default.zones[1].id
+  availability_zone = var.zone
   image_id          = data.ucloud_images.default.images[0].id
   instance_type     = "o-highcpu-2"
   root_password     = var.instance_password