+++ examples/web-server/variables.tf
@@ -1,6 +1,9 @@
 variable "region" {
-  description = "The region to create resources in"
-  default     = "cn-bj2"
+  default = "cn-bj2"
+}
+
+variable "zone" {
+  default = "cn-bj2-05"
 }
 
 variable "instance_password" {