+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## 1.22.0 (Unreleased)
+
+ENHANCEMENTS:
+
+* resource/ucloud_db_instance: update `instance_type` about NVMe SSD DB [GH-69]
+* resource/ucloud_instance: update `instance_type` about OS instance [GH-69]
+
 ## 1.21.0 (July 27, 2020)
 
 ENHANCEMENTS: